Move download_from_gcs to Starboard
Removes reverse dependency from Starboard to Cobalt

b/364421874
