+++ cobalt/renderer/BUILD.gn
@@ -500,7 +500,7 @@ copy("renderer_copy_lottie_test_data") {
 action("renderer_download_lottie_test_data") {
   install_content = true
 
-  script = "//cobalt/tools/download_from_gcs.py"
+  script = "//starboard/tools/download_from_gcs.py"
 
   inputs = [
     "$_lottie_resource_path/finger_print-expected.png.sha1",
