+++ components/crx_file/BUILD.gn
@@ -35,7 +35,7 @@ if (is_starboard) {
   action("crx_file_download_test_data") {
     install_content = true
 
-    script = "//cobalt/tools/download_from_gcs.py"
+    script = "//starboard/tools/download_from_gcs.py"
 
     sha_sources = []
     foreach(sha1_file, sha1_files) {
